# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Państwa Miasta - Game API
  description: |
    REST API for the "Państwa Miasta" (Countries & Cities) multiplayer word game.

    ## Game Flow
    1. Host creates a room → players join via room code
    2. Host configures settings (categories, time, rounds)
    3. Host starts the game → a random letter is drawn
    4. Players fill in answers for each category
    5. Players click "Send answers" (or time runs out) → transition to reviewing
    6. Players review and vote on each other's answers
    7. Host ends the round → scores are calculated → next round or finish

    ## Scoring
    - **15 pts** — only player with a valid answer in that category
    - **10 pts** — unique valid answer (different from others)
    - **5 pts** — duplicate valid answer (same as another player)
    - **0 pts** — empty or rejected by votes

    ## Persistence (PostgreSQL + CloudNativePG + JPA)
    The backend is a Spring Boot (Java 21) application using Spring Data JPA / Hibernate.
    Room state is persisted in PostgreSQL managed by **CloudNativePG** (`Cluster` `pm-postgres`, 3 instances,
    JDBC via `pm-postgres-rw` service). Schema: `rooms` table (primary key `code`, `varchar(10)`) and a
    `players` table (foreign key `room_code` → `rooms.code` with **ON DELETE CASCADE**, indexed by
    `idx_players_room_code`). Nested structures are stored as PostgreSQL `JSONB` columns.

    Connection parameters come from environment variables
...

...

...

...

...

...

...

...

...

    ```sql
...
    ```

...
servers:
...
paths:
  /rooms:
    get:
      tags: [Rooms]
      operationId: listPublicRooms
      summary: List public rooms
      description: Returns all public rooms that are currently in the lobby phase (waiting for players).
      responses:
        '200':
          description: |
            List of public rooms.
            Backed by `db.rooms.find({ isPublic: true, status: 'lobby' })`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicRoom'
              example:
                - code: "A1B2C3"
                  hostNick: "Wojtek"
                  playersCount: 3
                  maxPlayers: 8
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Rooms]
      operationId: createRoom
      summary: Create a new room
      description: Creates a new game room. The creator becomes the host.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '200':
          description: |
            Room created successfully. A new row is inserted into the `rooms` table
            with primary key `code` equal to the generated room code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResponse'
        '400':
          description: Validation error (e.g. missing nick)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Nick is required"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ws/rooms/{code}:
    get:
      tags: [Rooms]
      operationId: roomWebSocket
      summary: WebSocket room state stream
      description: |
        **WebSocket endpoint** (not HTTP GET). Connect to `ws(s)://<host>/api/ws/rooms/{code}`.

        **Client → server:**
        - `{ "type": "subscribe", "playerId": "<uuid>" }` — first message after connect
        - `{ "type": "ping" }` — heartbeat every 30s

        **Server → client:**
        - `{ "type": "room_state", "data": { ...Room... } }` — full room snapshot on subscribe and on every change
        - `{ "type": "pong" }` — response to ping
        - `{ "type": "error", "message": "..." }` — e.g. invalid playerId

        Mutations (start, stop, vote, …) use REST endpoints; state changes are pushed via Redis Pub/Sub
        to all backend replicas.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '101':
          description: Switching Protocols (WebSocket upgrade)
  /rooms/{code}:
    get:
      tags: [Rooms]
      operationId: getRoomState
      summary: Get room state (REST fallback)
      description: |
        Returns the full state of the room including players, settings, game state, and scores.
        Primary sync is via WebSocket (`/api/ws/rooms/{code}`); this endpoint is used as fallback
        on reconnect and for debugging.

        When the room is in `playing` status, the response includes `mainTimeLeft`
        (seconds remaining in the round) computed directly from the persisted `game.roundEndsAt`
        field: `max(0, floor((roundEndsAt - now) / 1000))`.

        Before loading the room, the endpoint runs the atomic auto-stop `UPDATE` that transitions
        it to `reviewing` if `round_ends_at <= now` (see the "Scalable auto-stop" section).

        Implemented as `roomRepository.findById(code)` after the optional auto-stop update.
        The row is returned as the JSON room object plus the computed `mainTimeLeft`.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '200':
          description: Full room state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomState'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/join:
    post:
      tags: [Rooms]
      operationId: joinRoom
      summary: Join a room
      description: |
        Join an existing room by its code. If a player with the same nick already exists
        in the room, the existing player ID is returned (reconnect behavior).

        Performs one read and (on success) one save against the `rooms` table.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomRequest'
      responses:
        '200':
          description: Successfully joined (or reconnected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResponse'
        '400':
          description: Game already started, room is full, or nick missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/settings:
    post:
      tags: [Settings]
      operationId: updateSettings
      summary: Update room settings (host only)
      description: |
        Updates the room settings. Only the host can perform this action. 
        Settings can only be changed while the room is in `lobby` status.
        Supports partial updates — only provided fields are overwritten.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated (room row updated in PostgreSQL)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cannot change settings (game not in lobby)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only host can change settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/start:
    post:
      tags: [Gameplay]
      operationId: startGame
      summary: Start the game / next round
      description: |
        Starts the game (or next round after reviewing). Only the host can start the game.
        A random letter is drawn, the round timer begins, and when time expires
        the status automatically transitions to `reviewing`.

        On the first round, all player scores are reset to 0.

        The room row in PostgreSQL is updated with the new `status`, `currentRound`,
        `currentLetter`, `roundStartedAt`, `roundEndsAt` and cleared `answers`/`votes`/`stoppedPlayers`.
        `roundEndsAt` is set to `roundStartedAt + settings.timePerRound * 60 000` ms and is used
        by the polling endpoint for atomic auto-stop (no in-process timers).
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerIdRequest'
      responses:
        '200':
          description: Game started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cannot start game (wrong status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only host can start game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/stop:
    post:
      tags: [Gameplay]
      operationId: triggerStop
      summary: Trigger STOP
      description: |
        A player signals that they want to stop the round. Each player can trigger stop once.
        When **all** players have triggered stop, the status automatically transitions to `reviewing`.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerIdRequest'
      responses:
        '200':
          description: |
            Stop registered. The player id is appended to `game.stoppedPlayers`
            in the persisted room document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Not in playing phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Player not in room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/answers:
    post:
      tags: [Gameplay]
      operationId: submitAnswers
      summary: Submit answers
      description: |
        Submit the player's answers for the current round. 
        Answers are a map of category name → answer string.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswersRequest'
            example:
              playerId: "550e8400-e29b-41d4-a716-446655440000"
              answers:
                Państwa: "Polska"
                Miasta: "Poznań"
                Zwierzęta: "Panda"
                Rośliny: "Paproć"
                Imiona: "Patrycja"
      responses:
        '200':
          description: |
            Answers submitted. The player's answers map is stored under
            `game.answers[playerId]` in the room document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/vote:
    post:
      tags: [Gameplay]
      operationId: submitVote
      summary: Submit a vote
      description: |
        Submit a vote on another player's answer for a specific category.
        Voting is only possible during the `reviewing` phase.
        
        Each voter can cast one vote per (target player, category) pair.
        A vote of `isValid: true` means the answer is accepted; `false` means rejected.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitVoteRequest'
      responses:
        '200':
          description: |
            Vote recorded under `game.votes[targetPlayerId][category][voterId]`
            in the persisted room document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Not in reviewing phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Voter not in room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/next-round:
    post:
      tags: [Gameplay]
      operationId: nextRound
      summary: End round / go to next round (host only)
      description: |
        Ends the current reviewing phase and either starts the next round 
        or finishes the game if all rounds have been played.
        
        **Scoring is calculated** during this transition:
        - 15 pts if the player is the only one with a valid answer in that category
        - 10 pts for a unique valid answer
        - 5 pts for a duplicate valid answer
        - 0 pts if the answer was empty or rejected by majority vote
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerIdRequest'
      responses:
        '200':
          description: |
            Round ended; the next round is started or the game is marked `finished`.
            Updated `scores`, `currentRound`, `currentLetter`, `roundStartedAt`, `roundEndsAt`,
            and cleared `answers`/`votes`/`stoppedPlayers` are persisted to PostgreSQL.
            `roundEndsAt` is set to `null` when the game finishes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Not in reviewing phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only host can end the round
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/leave:
    post:
      tags: [Rooms]
      operationId: leaveRoom
      summary: Leave room
      description: |
        Removes the player from the room in any phase (lobby, playing, reviewing, finished).
        Host role transfers to the next player if others remain. If the leaving player is the
        last member, the room is deleted immediately from PostgreSQL and Redis keys are cleaned up.
        Otherwise remaining clients receive a WebSocket push with updated state.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerIdRequest'
      responses:
        '200':
          description: Player left successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Player not in room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /rooms/{code}/reset:
    post:
      tags: [Gameplay]
      operationId: resetToLobby
      summary: Reset to lobby (host only)
      description: |
        Resets the room back to the lobby state. All game progress and scores are cleared.
        Only the host can perform this action.
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerIdRequest'
      responses:
        '200':
          description: |
            Room reset to lobby. All in-progress game fields are cleared and
            scores are zeroed in the persisted row. The room itself is **not**
            removed from PostgreSQL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Only host can reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Player session JWT from create/join (`accessToken`)
  parameters:
    RoomCode:
      name: code
      in: path
      required: true
      description: 6-character alphanumeric room code
      schema:
        type: string
        pattern: '^[A-Z0-9]{6}$'
        example: "A1B2C3"
  responses:
    InternalServerError:
      description: |
        Unexpected backend error. In practice this is most often a PostgreSQL
        connectivity or query failure (the backend logs the original error and
        returns a generic message to the client).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
    NotFound:
      description: Room does not exist in the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Room not found"
  schemas:
...
    CreateRoomRequest:
      type: object
      required: [nick]
      properties:
        nick:
          type: string
          description: Nickname of the player creating the room (becomes host)
          example: "Wojtek"
        isPublic:
          type: boolean
          description: Whether the room is visible in the public room list
          default: false
          example: true
    JoinRoomRequest:
      type: object
      required: [nick]
      properties:
        nick:
          type: string
          description: Nickname of the joining player
          example: "Anna"
    PlayerIdRequest:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID of the player performing the action
          example: "550e8400-e29b-41d4-a716-446655440000"
    UpdateSettingsRequest:
      type: object
      required: [playerId, settings]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID of the host player
          example: "550e8400-e29b-41d4-a716-446655440000"
        settings:
          $ref: '#/components/schemas/RoomSettings'
    SubmitAnswersRequest:
      type: object
      required: [playerId, answers]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID of the player submitting answers
          example: "550e8400-e29b-41d4-a716-446655440000"
        answers:
          type: object
          additionalProperties:
            type: string
          description: "Map of category name → player's answer"
          example:
            Państwa: "Polska"
            Miasta: "Poznań"
            Zwierzęta: "Panda"
    SubmitVoteRequest:
      type: object
      required: [voterId, targetPlayerId, category, isValid]
      properties:
        voterId:
          type: string
          format: uuid
          description: UUID of the player casting the vote
          example: "550e8400-e29b-41d4-a716-446655440000"
        targetPlayerId:
          type: string
          format: uuid
          description: UUID of the player whose answer is being voted on
          example: "661e8400-e29b-41d4-a716-446655440001"
        category:
          type: string
          description: Category of the answer being voted on
          example: "Miasta"
        isValid:
          type: boolean
          description: Whether the voter considers the answer valid
          example: true

    # ── Response Schemas ─────────────────────────────────────
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Room not found"
    JoinResponse:
      type: object
      properties:
        code:
          type: string
          description: Room code
          example: "A1B2C3"
        playerId:
          type: string
          format: uuid
          description: UUID assigned to the player
          example: "550e8400-e29b-41d4-a716-446655440000"
        accessToken:
          type: string
          description: Bearer JWT for subsequent REST calls and WebSocket subscribe
    PublicRoom:
      type: object
      properties:
        code:
          type: string
          description: Room code
          example: "A1B2C3"
        hostNick:
          type: string
          description: Nickname of the room host
          example: "Wojtek"
        playersCount:
          type: integer
          description: Current number of players in the room
          example: 3
        maxPlayers:
          type: integer
          description: Maximum players allowed
          example: 8

    # ── Data Model Schemas ───────────────────────────────────
    Player:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique player identifier
        nick:
          type: string
          description: Player's nickname
        isHost:
          type: boolean
          description: Whether this player is the room host
    RoomSettings:
      type: object
      description: Configurable game settings (all fields optional for partial updates)
      properties:
        categories:
          type: array
          items:
            type: string
          description: List of category names for each round
          example: ["Państwa", "Miasta", "Zwierzęta", "Rośliny", "Imiona"]
        timePerRound:
          type: integer
          description: Time per round in minutes
          minimum: 1
          example: 1
        rounds:
          type: integer
          description: Total number of rounds
          minimum: 1
          example: 5
        maxPlayers:
          type: integer
          description: Maximum number of players
          minimum: 2
          maximum: 20
          example: 8
...
    RoomState:
      type: object
      description: |
        Complete room state returned by the polling endpoint.

        Persisted as a row in the `rooms` table (primary key `code`) in PostgreSQL;
        nested structures are stored as `JSONB` columns.
      properties:
        code:
          type: string
          description: Room code (primary key)
          example: "A1B2C3"
        status:
          type: string
          enum: [lobby, playing, reviewing, finished]
          description: |
            Current phase of the room:
            - `lobby` — waiting for players, settings can be changed
            - `playing` — round in progress, players fill in answers
            - `reviewing` — players vote on answers
            - `finished` — all rounds completed, final scores shown
          example: "playing"
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
        scores:
          type: object
          additionalProperties:
            type: integer
          description: "Map of playerId → cumulative score"
          example:
            "550e8400-e29b-41d4-a716-446655440000": 35
            "661e8400-e29b-41d4-a716-446655440001": 20
        settings:
          $ref: '#/components/schemas/RoomSettings'
        game:
          $ref: '#/components/schemas/GameState'